+++ src/main/java/com/kustacks/kuring/worker/scrap/client/staff/KuStaffApiClient.java
@@ -3,6 +3,7 @@ package com.kustacks.kuring.worker.scrap.client.staff;
 import com.kustacks.kuring.common.error.ErrorCode;
 import com.kustacks.kuring.common.error.InternalLogicException;
 import com.kustacks.kuring.notice.domain.DepartmentName;
+import com.kustacks.kuring.worker.scrap.client.JsoupClient;
 import com.kustacks.kuring.worker.scrap.deptinfo.DeptInfo;
 import com.kustacks.kuring.worker.scrap.deptinfo.art_design.CommunicationDesignDept;
 import com.kustacks.kuring.worker.scrap.deptinfo.art_design.LivingDesignDept;
