+++ src/main/java/com/kustacks/kuring/worker/scrap/client/staff/RealEstateStaffApiClient.java
@@ -2,6 +2,7 @@ package com.kustacks.kuring.worker.scrap.client.staff;
 
 import com.kustacks.kuring.common.error.ErrorCode;
 import com.kustacks.kuring.common.error.InternalLogicException;
+import com.kustacks.kuring.worker.scrap.client.JsoupClient;
 import com.kustacks.kuring.worker.scrap.deptinfo.DeptInfo;
 import com.kustacks.kuring.worker.scrap.deptinfo.real_estate.RealEstateDept;
 import org.jsoup.nodes.Document;
